Fewer Gathers and Variable Number of Writers
Reduced the number of gathers required for writing to one for each writer. Also introduced the
ability to specify the number of writers per branch as a command line argument. The hdf5 to vtu
conversion scripts have been modified to work with the new system. They now require two extra
parameters - the number of writers per branch and the number of quads each writer writes for.