Clean IP3 Model Code And Move Plotting Script To Util
The (modified) Lemon model now uses a constant PIP2 concentration.

Plotting cannot occur if a combination of explicit and implicit solving is set
(instead of creating an empty csv file and failing before).

Local plotting script has been tidied and moved to util folder. Expected arguments
is the csv file name/path, the model used ('o', 'l', 'b', for the original, lemon,
or bennett model respectively). An optional colour argument can be provided in the
form of a hexidecimal string with a preceeding "-c".